+++ CoffeShop/CoffeeShop.DataAccess/Repositories/CustomRepositories/OrderRepositories/OrderRepository.cs
@@ -7,7 +7,7 @@ public class OrderRepository : Repository<Order>, IOrderRepository
 {
     public OrderRepository(CoffeeShopContext context) : base(context) { }
 
-    public async Task<Order> GetAsync(int id)
+    public async Task<Order> GetAsync(long id)
         => await _context.Orders.FirstOrDefaultAsync(x => x.Id.Equals(id))
            ?? throw new NullReferenceException();
 }
