+++ CoffeShop/CoffeeShop.DataAccess/Repositories/CustomRepositories/VolumeRepositories/VolumeRepository.cs
@@ -7,7 +7,7 @@ public class VolumeRepository : Repository<Volume>, IVolumeRepository
 {
     public VolumeRepository(CoffeeShopContext context) : base(context) { }
 
-    public async Task<Volume> GetAsync(int id)
+    public async Task<Volume> GetAsync(long id)
         => await _context.Volumes.FirstOrDefaultAsync(x=>x.Id.Equals(id)) 
            ?? throw new NullReferenceException();
 }
